% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adduct_rules.R
\name{adduct_rules}
\alias{adduct_rules}
\title{Create list with all adduct calculation rules (all rules)

This function returns a list with rules for the calculation of adducts. It is required for the calculation of adduct m/z values from exact masses and the other way round. This list contains all rules for both ionization modes.}
\usage{
adduct_rules()
}
\description{
Create list with all adduct calculation rules (all rules)

This function returns a list with rules for the calculation of adducts. It is required for the calculation of adduct m/z values from exact masses and the other way round. This list contains all rules for both ionization modes.
}
\seealso{
\code{\link{adduct_rules_pos}}

\code{\link{adduct_rules_neg}}

\code{\link{get_adduct_names}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
