% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adduct_calc.R
\name{calc_adduct_mass}
\alias{calc_adduct_mass}
\title{Calculate adduct mass from neutral exact mass}
\usage{
calc_adduct_mass(exact_mass, adduct)
}
\description{
Calculate adduct mass from neutral exact mass
}
