% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_formula_math.R
\name{contains_formula}
\alias{contains_formula}
\title{This function checks if one sum formula is contained in another.}
\usage{
contains_formula(target_chem_formula, query_chem_formula)
}
\arguments{
\item{target_formula}{Single string with chemical formula}

\item{query_formula}{Single string with chemical formula that should be contained in the targetFormula}
}
\value{
TRUE or FALSE
}
\description{
This function checks if one sum formula is contained in another.
}
\examples{
contains_formula("C6H12O6", "H2O")
}
