% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_formula_parser.R
\name{formula_to_list}
\alias{formula_to_list}
\title{This function parses a chemical chem_formula into a named vector}
\usage{
formula_to_list(chem_formula)
}
\arguments{
\item{chem_formula}{Single string with chemical chem_formula}
}
\value{
Named vector with all elements
}
\description{
This function parses a chemical chem_formula into a named vector
}
\examples{
parseChemFormula("C6H12O6")
}
