% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_formula_parser.R
\name{list_to_formula}
\alias{list_to_formula}
\title{This function generates a chemical chem_formula from a named vector of elemental counts}
\usage{
list_to_formula(formula_list)
}
\arguments{
\item{formula_list}{Named list containing the elements as names and their abudance as values}
}
\value{
Single string with chemical chem_formula
}
\description{
This function generates a chemical chem_formula from a named vector of elemental counts
}
\examples{
chem_formula <- list("C" = 6,
                "H" = 12,
                "O" = 6)
generateformula(chem_formula)
}
