% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_mona_json.R
\name{mona_isolate_id}
\alias{mona_isolate_id}
\title{Get the ID of a MoNA entry

This function returns the ID associated with a specific entry in a MoNA JSON object}
\usage{
mona_isolate_id(x)
}
\arguments{
\item{x}{A single entry of a list from a parsed MoNA JSON}
}
\description{
Get the ID of a MoNA entry

This function returns the ID associated with a specific entry in a MoNA JSON object
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
