% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_formula_parser.R
\name{standardize_formula}
\alias{standardize_formula}
\title{This function standardizes a supplied chemical chem_formula according to the Hill notation system.}
\usage{
standardize_formula(chem_formula)
}
\arguments{
\item{chem_formula}{Single string with chemical chem_formula}
}
\value{
Single string with standardized chemical formua
}
\description{
This function standardizes a supplied chemical chem_formula according to the Hill notation system.
}
\examples{
standardizeChemFormula("O6C6H12")
}
